feat(availability): pr-269 let-patient-choose-address and drawer views restructure
- add letPatientChooseAddress boolean to ISlot (BE types, Mongoose schema,
  editSlotHandler, AvailabilityService.editSlot)
- add letPatientChooseAddress to FE ISlot, IWeekSlot, editSlot API call,
  useSlotAddress hook (with dedicated mutation + PostHog event)
- add SlotLocationSection card with three-toggle progressive disclosure UX
  (let patient choose → share clinic → override address)
- restructure all drawer views into dedicated kebab-case folders, each with
  a co-located .types.ts file
- move AvailabilityExtendBanner to drawer/extend-banner/
- move useWeekSlots to week/hook/
