+++ src/pages/availability/week/drawer/extend-banner/AvailabilityExtendBanner.tsx
