+++ src/pages/availability/week/drawer/hooks/useSlotAddress.ts
@@ -1,5 +1,7 @@
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { capture } from '@psycron/analytics/posthog/events';
+import { PostHogEvent } from '@psycron/analytics/posthog/types';
 import { editSlot } from '@psycron/api/availability';
 import { useAlert } from '@psycron/context/alert/AlertContext';
 import type { ISlotAddress } from '@psycron/context/user/auth/UserAuthenticationContext.types';
@@ -17,6 +19,10 @@ export const useSlotAddress = (
 
 	const initialAddress = slot.address ?? null;
 	const [address, setAddress] = useState<ISlotAddress | null>(initialAddress);
+
+	const [letPatientChoose, setLetPatientChooseState] = useState(
+		slot.letPatientChooseAddress ?? false
+	);
 
 	const isDirty =
 		JSON.stringify(address) !== JSON.stringify(initialAddress);
@@ -44,7 +50,42 @@ export const useSlotAddress = (
 		},
 	});
 
+	const letPatientChooseMutation = useMutation({
+		mutationFn: (val: boolean) =>
+			editSlot({
+				availabilityDayId: slot.availabilityDayId ?? '',
+				letPatientChooseAddress: val,
+				slotId: slot._id ?? slot.id,
+				therapistId: therapistId ?? '',
+			}),
+		onError: () => {
+			showAlert({
+				message: t('availability.week.drawer.address-save-error'),
+				severity: 'error',
+			});
+		},
+		onSuccess: (_, val) => {
+			capture(PostHogEvent.AvailabilityLetPatientChooseAddress, {
+				enabled: val,
+			});
+			queryClient.invalidateQueries({ queryKey: ['therapistAvailability'] });
+		},
+	});
+
+	const setLetPatientChoose = (val: boolean) => {
+		setLetPatientChooseState(val);
+		letPatientChooseMutation.mutate(val);
+	};
+
 	const clear = () => setAddress(null);
 
-	return { address, clear, isDirty, mutation, setAddress };
+	return {
+		address,
+		clear,
+		isDirty,
+		letPatientChoose,
+		mutation,
+		setAddress,
+		setLetPatientChoose,
+	};
 };
