+++ src/pages/availability/week/drawer/views/slot-available-body/SlotAvailableBody.types.ts
@@ -1,0 +1,9 @@
+import type { UseFormReturn } from 'react-hook-form';
+import type { ICreatePatientForm } from '@psycron/api/patient/index.types';
+
+import type { ISlotLocationSectionProps } from '../slot-location-section/SlotLocationSection.types';
+
+export interface ISlotAvailableBodyProps extends ISlotLocationSectionProps {
+	methods: UseFormReturn<ICreatePatientForm>;
+	showSessionLocation: boolean;
+}
