+++ src/pages/availability/week/drawer/views/slot-cancel-choice-view/SlotCancelChoiceView.types.ts
@@ -1,0 +1,4 @@
+export interface ISlotCancelChoiceViewProps {
+	onCancel: () => void;
+	onReschedule: () => void;
+}
