+++ src/pages/availability/week/drawer/views/slot-cancel-reason-form/SlotCancelReasonForm.types.ts
@@ -1,0 +1,8 @@
+import type { CancellationReasonEnum as CancellationReasonType } from '@psycron/api/user/availability/index.types';
+
+export interface ISlotCancelReasonFormProps {
+	customReason: string;
+	onCustomReasonChange: (val: string) => void;
+	onReasonChange: (val: CancellationReasonType) => void;
+	reasonCode: CancellationReasonType | null;
+}
