+++ src/pages/availability/week/drawer/views/slot-detail-view/SlotDetailView.types.ts
@@ -1,0 +1,5 @@
+import type { IDrawerDetail } from '../../AvailabilityWeekDrawer.types';
+
+export interface ISlotDetailViewProps {
+	details: IDrawerDetail[];
+}
