+++ src/pages/availability/week/drawer/views/slot-reschedule-picker/SlotReschedulePicker.types.ts
@@ -1,0 +1,14 @@
+import type { IRescheduleSlot } from '../../AvailabilityWeekDrawer.types';
+
+export interface ISlotGroup {
+	availabilityDayId: string;
+	date: string;
+	formattedDate: string;
+	slots: IRescheduleSlot[];
+}
+
+export interface ISlotReschedulePickerProps {
+	availableSlotGroups: ISlotGroup[];
+	onSelectSlot: (slot: IRescheduleSlot) => void;
+	selectedSlot: IRescheduleSlot | null;
+}
